requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.3
pillow==10.0.0
gunicorn==21.2.0